fix: four Advanced Heading and Text Path defects
**Typing hid the last letter of every word.** The effect ran on
steps(12, end), which reaches its final value only at exactly t=1 --
and an infinite animation restarts at that instant, so the fully
revealed frame was never painted. Sampling the computed clip-path
across a cycle showed twelve values from 100% down to 8.33% and never
0%. It now uses jump-none, which paints both ends of the range, and
takes its step count from the word's character count instead of a
hardcoded 12, so a step reveals a letter rather than a fixed fraction
of the width.

**The clip-path effects shaved ascenders and descenders.** inset()
clips to the border box, which for an inline-block is the line box.
Measured on a rendered heading: the box is 36px while the glyph ink of
'gjpqyA' is 38.9px, so 2.9px was cut. typing, clip and blinds now
bleed -0.3em vertically, which expands the clip rect past the border
box.

**Text Path motion had no editor preview.** It works on the frontend --
verified by watching startOffset advance on a published page -- but
view.js is a viewScript the editor never loads, so switching Motion on
did nothing on the canvas and read as broken. The engine moves to
motion.js, shared by view.js and edit.js, and the editor starts and
stops it as the motion attributes change. getOpenPathMaximumOffset()
now resolves the guide through the textPath's own ownerDocument,
because in the editor the block lives in the canvas iframe and a bare
document lookup searches the parent frame.

**Plain heading segments were squished in the editor.** The frontend
gap does not come from the stylesheet -- the serializer writes newlines
between the segment blocks and that whitespace collapses to a single
space. The editor renders the segments adjacent with no text node
between them. Animated segments already carried an editor-only
compensating gap; plain ones never got one. A segment renders either
__animated or __text and never both, so mirroring the rule onto __text
cannot double up. The variable is renamed --dsgo-editor-segment-gap
now that it serves both.
